+++ UI/GameEditorLibrary/Controls/ValueEditors/SpriteAnimationEditor.xaml.cs
@@ -22,12 +22,16 @@
         private readonly IProjectService _projectService = ViewContainer.Resolve<IProjectService>();
 
         public SpriteAnimationEditor() : base() {
-            this.SelectSpriteAnimationCommand = new RelayCommand(() => {
+            this.SelectCommand = new RelayCommand(() => {
                 if (this._dialogService.ShowSelectAssetDialog(this._projectService.CurrentProject, typeof(SpriteAnimationAsset), true, out var asset)) {
                     this.Asset = asset as SpriteAnimationAsset;
                     this.Value = this.Asset?.SavableValue;
                 }
             }, true);
+
+            this.ClearCommand = new RelayCommand(() => {
+                this.Value = null;
+            });
 
             this.Loaded += this.SpriteAnimationEditor_Loaded;
             this.InitializeComponent();
@@ -38,7 +42,9 @@
             set { this.SetValue(AssetProperty, value); }
         }
 
-        public ICommand SelectSpriteAnimationCommand { get; }
+        public ICommand ClearCommand { get; }
+
+        public ICommand SelectCommand { get; }
 
         private void SpriteAnimationEditor_Loaded(object sender, RoutedEventArgs e) {
             if (this.Value != null && (this.Asset == null || this.Asset.SavableValue?.Id != this.Value.Id)) {